Add YAML support, handler implementation
Further information on this to be written.

YAML is added as an alternative to JSON (the two are independently
supported because they're incompatible) because it's a little
unfortunate having to type a lot of string literals and escape them
inside of gojq expressions. YAML's multi-line string literals are
a minor improvement over this, but have the trade-off of being YAML.